CNN/BovinaBreedPrediction/gcp/requirements.txt
tensorflow==2.20.0
numpy
Pillow
google-cloud-storage
python-multipart
fastapi
uvicorn